requirements.txt
Flask
Flask-SQLAlchemy
pymysql
mysql-connector-python

